chore: 친구 추가하기 페이지 레이아웃 수정
